refactor: InstantTaskExecutorExtension.kt 패키지 이동
- 이전: staccato/presentation
- 이후: staccato
