+++ android/Staccato_AN/app/src/test/java/com/on/staccato/presentation/category/viewmodel/CategoryViewModelTest.kt
@@ -1,12 +1,12 @@
 package com.on.staccato.presentation.category.viewmodel
 
 import com.on.staccato.CoroutinesTestExtension
+import com.on.staccato.InstantTaskExecutorExtension
 import com.on.staccato.data.Exception
 import com.on.staccato.data.ServerError
 import com.on.staccato.data.Success
 import com.on.staccato.data.dto.Status
 import com.on.staccato.domain.repository.CategoryRepository
-import com.on.staccato.presentation.InstantTaskExecutorExtension
 import com.on.staccato.presentation.category.INVALID_ID
 import com.on.staccato.presentation.category.VALID_ID
 import com.on.staccato.presentation.category.category
